also move out samples to FeatureExtractor- and ...SubjectChangerSamples
moreover, use toBeEmpty naming also in infix, this way it is easier to
recognise the same function in api-fluent and one thing less to adopt
when copying samples from api-fluent
