+++ apis/fluent-en_GB/atrium-api-fluent-en_GB-common/src/test/kotlin/ch/tutteli/atrium/api/fluent/en_GB/samples/Fun0ExpectationSamples.kt
@@ -24,13 +24,20 @@ class Fun0ExpectationSamples {
         expect { throw IllegalStateException("abc") }
             .toThrow<IllegalStateException> { // subject inside this block is of type IllegalStateException
                 messageToContain("abc")
-            }
+            } // subject keeps type IllegalStateException also after the block
 
         fails { // because wrong type expected (IndexOutOfBoundsException instead of IllegalStateException), but since we use a block...
             expect { throw IllegalStateException("abc") }
                 .toThrow<IndexOutOfBoundsException> {
                     messageToContain("abc") // ... reporting mentions that subject's message was expected `to contain: "abc"`
                 }
+        }
+
+        fails {
+            // because you forgot to define an expectation in the expectation group block
+            // use `.toThrow()` if this is all you expect
+            expect { throw IllegalStateException("abc") }
+                .toThrow<IndexOutOfBoundsException> {}
         }
     }
 
@@ -51,11 +58,26 @@ class Fun0ExpectationSamples {
         expect { "abc" }
             .notToThrow { // subject is now of type String
                 toEqual("abc")
-            }
+            } // subject keeps type String also after the block
+
+        fails { // because an exception was thrown, but since we use a block...
+            expect<() -> String> { throw IllegalStateException("abc") }
+                .notToThrow {
+                    toStartWith("abc") // ... reporting mentions that subject's message was expected `to start with: "abc"`
+                }
+        }
+
+        fails { // because an exception was thrown, but since we use a block...
+            expect<() -> String> { throw IllegalStateException("abc") }
+                .notToThrow {
+                    toStartWith("abc") // ... reporting mentions that subject's message was expected `to start with: "abc"`
+                }
+        }
 
         fails {
-            expect { throw IllegalStateException("abc") }
-                .notToThrow {}
+            // because you forgot to define an expectation in the expectation group block
+            // use `.notToThrow()` if this is all you expect
+            expect { "abc" }.notToThrow {}
         }
     }
 }
