+++ apis/infix-en_GB/atrium-api-infix-en_GB-common/src/main/kotlin/ch/tutteli/atrium/api/infix/en_GB/mapExpectations.kt
@@ -258,26 +258,3 @@ infix fun <T : Map<*, *>> Expect<T>.notToBe(@Suppress("UNUSED_PARAMETER") empty:
     _logicAppend { isNotEmpty(::toEntries) }
 
 private fun <T : Map<*, *>> toEntries(t: T): Collection<*> = t.entries
-
-/**
- * Creates an [Expect] for the property [Collection.size] of the subject of `this` expectation,
- * so that further fluent calls are assertions about it.
- *
- * @return The newly created [Expect] for the extracted feature.
- *
- * @since 0.15.0
- */
-val <T : Map<*, *>> Expect<T>.size: Expect<Int>
-    get() = _logic.size(::toEntries).transform()
-
-/**
- * Expects that the property [Collection.size] of the subject of `this` expectation
- * holds all assertions the given [assertionCreator] creates for it and
- * returns an [Expect] for the current subject of `this` expectation.
- *
- * @return an [Expect] for the subject of `this` expectation.
- *
- * @since 0.15.0
- */
-infix fun <K, V, T : Map<out K, V>> Expect<T>.size(assertionCreator: Expect<Int>.() -> Unit): Expect<T> =
-    _logic.size(::toEntries).collectAndAppend(assertionCreator)
