+++ apis/infix-en_GB/atrium-api-infix-en_GB-common/src/test/kotlin/ch/tutteli/atrium/api/infix/en_GB/samples/ListFeatureExtractorSamples.kt
@@ -5,7 +5,7 @@ import ch.tutteli.atrium.api.verbs.internal.expect
 import ch.tutteli.atrium.translations.DescriptionComparableAssertion
 import kotlin.test.Test
 
-class ListExpectationSamples {
+class ListFeatureExtractorSamples {
     private val toBeLessThanDescr = DescriptionComparableAssertion.IS_LESS_THAN.getDefault()
     private val toBeGreaterThanDescr = DescriptionComparableAssertion.IS_GREATER_THAN.getDefault()
